# xadminapi/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "adminapi",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api/v1",
    "paths": {
...
        "/admin_log/get_login_log": {
            "get": {
                "tags": [
                    "后台日志"
                ],
                "summary": "获取登录日志",
                "parameters": [
                    {
                        "type": "string",
                        "description": "token",
                        "name": "x-token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "操作人",
                        "name": "account",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "渠道商",
                        "name": "channel_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "结束时间",
                        "name": "end_time",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "登录Ip",
                        "name": "login_ip",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "页码",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "每页数量",
                        "name": "page_size",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "运营商",
                        "name": "seller_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "开始时间",
                        "name": "start_time",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.XAdminLoginLog"
                            }
                        }
                    }
                }
            }
        },
        "/admin_log/get_opt_log": {
            "get": {
                "tags": [
                    "后台日志"
                ],
                "summary": "获取操作日志",
                "parameters": [
                    {
                        "type": "string",
                        "description": "操作人",
                        "name": "account",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "渠道商",
                        "name": "channel_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "结束时间",
                        "name": "end_time",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "操作名",
                        "name": "opt_name",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "页码",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "每页数量",
                        "name": "page_size",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "运营商",
                        "name": "seller_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "开始时间",
                        "name": "start_time",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.XAdminOptLog"
                            }
                        }
                    }
                }
            }
        },
...
        "/channel/create_channel": {
            "post": {
                "tags": [
                    "渠道"
                ],
                "summary": "新增渠道",
                "parameters": [
                    {
                        "type": "string",
                        "description": "token",
                        "name": "x-token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "验证码",
                        "name": "VerifyCode",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "body参数",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/service_admin.CreateXChannelReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功"
                    }
                }
            }
        },
        "/channel/delete_channel": {
            "delete": {
                "tags": [
                    "渠道"
                ],
                "summary": "删除渠道",
                "parameters": [
                    {
                        "type": "string",
                        "description": "token",
                        "name": "x-token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "验证码",
                        "name": "VerifyCode",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "body参数",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/service_admin.DeleteXChannelReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功"
                    }
                }
            }
        },
        "/channel/get_channel": {
            "get": {
                "tags": [
                    "渠道"
                ],
                "summary": "获取渠道列表",
                "parameters": [
                    {
                        "type": "string",
                        "description": "token",
                        "name": "x-token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "渠道商",
                        "name": "channel_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "渠道商名称",
                        "name": "channel_name",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "页码",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "每页数量",
                        "name": "page_size",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "运营商",
                        "name": "seller_id",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.XChannel"
                            }
                        }
                    }
                }
            }
        },
        "/channel/update_channel": {
            "patch": {
                "tags": [
                    "渠道"
                ],
                "summary": "更新渠道",
                "parameters": [
                    {
                        "type": "string",
                        "description": "token",
                        "name": "x-token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "验证码",
                        "name": "VerifyCode",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "body参数",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/service_admin.UpdateXChannelReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功"
                    }
                }
            }
        },
        "/config/get_config": {
            "post": {
                "tags": [
                    "系统设置"
                ],
                "summary": "获取配置",
                "parameters": [
                    {
                        "type": "string",
                        "description": "token",
                        "name": "x-token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "筛选参数",
                        "name": "body",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/service_admin.GetXConfigReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "$ref": "#/definitions/service_admin.GetXConfigRes"
                        }
                    }
                }
            }
        },
        "/config/update_config": {
            "patch": {
                "tags": [
                    "系统设置"
                ],
                "summary": "更新配置",
                "parameters": [
                    {
                        "description": "筛选参数",
                        "name": "body",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/service_admin.GetXConfigReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功"
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "model.XAdminLoginLog": {
            "type": "object",
            "properties": {
                "account": {
                    "description": "账号",
                    "type": "string"
                },
                "channel_id": {
                    "description": "渠道商",
                    "type": "integer"
                },
                "create_at": {
                    "description": "创建时间",
                    "type": "string"
                },
                "id": {
                    "description": "自增Id",
                    "type": "integer"
                },
                "ip_location": {
                    "description": "ip地理位置",
                    "type": "string"
                },
                "login_ip": {
                    "description": "最近一次登录Ip",
                    "type": "string"
                },
                "memo": {
                    "description": "备注",
                    "type": "string"
                },
                "seller_id": {
                    "description": "运营商",
                    "type": "integer"
                },
                "update_at": {
                    "description": "修改时间",
                    "type": "string"
                }
            }
        },
        "model.XAdminOptLog": {
            "type": "object",
            "properties": {
                "account": {
                    "description": "账号",
                    "type": "string"
                },
                "channel_id": {
                    "description": "渠道商",
                    "type": "integer"
                },
                "create_at": {
                    "description": "创建时间",
                    "type": "string"
                },
                "id": {
                    "description": "自增Id",
                    "type": "integer"
                },
                "ip_location": {
                    "description": "Ip地理位置",
                    "type": "string"
                },
                "opt_name": {
                    "description": "请求路径",
                    "type": "string"
                },
                "req_data": {
                    "description": "请求参数",
                    "type": "string"
                },
                "req_ip": {
                    "description": "请求的Ip",
                    "type": "string"
                },
                "req_path": {
                    "description": "请求路径",
                    "type": "string"
                },
                "seller_id": {
                    "description": "运营商",
                    "type": "integer"
                },
                "update_at": {
                    "description": "更新时间",
                    "type": "string"
                }
            }
        },
...
        "model.XChannel": {
            "type": "object",
            "properties": {
                "channel_id": {
                    "description": "渠道商ID",
                    "type": "integer"
                },
                "channel_name": {
                    "description": "渠道名称",
                    "type": "string"
                },
                "create_at": {
                    "description": "创建时间",
                    "type": "string"
                },
                "id": {
                    "description": "Id",
                    "type": "integer"
                },
                "memo": {
                    "description": "备注",
                    "type": "string"
                },
                "seller_id": {
                    "description": "运营商ID",
                    "type": "integer"
                },
                "state": {
                    "description": "状态 1开启 2关闭",
                    "type": "integer"
                },
                "update_at": {
                    "description": "更新时间",
                    "type": "string"
                }
            }
        },
        "model.XConfig": {
            "type": "object",
            "properties": {
                "channel_id": {
                    "description": "渠道Id",
                    "type": "integer"
                },
                "config_name": {
                    "description": "配置名称",
                    "type": "string"
                },
                "config_value": {
                    "description": "配置值",
                    "type": "string"
                },
                "create_at": {
                    "description": "创建时间",
                    "type": "string"
                },
                "id": {
                    "description": "配置Id",
                    "type": "integer"
                },
                "memo": {
                    "description": "备注",
                    "type": "string"
                },
                "seller_id": {
                    "description": "运营商Id",
                    "type": "integer"
                },
                "update_at": {
                    "description": "更新时间",
                    "type": "string"
                }
            }
        },
...
        "service_admin.CreateXChannelReq": {
            "type": "object",
            "required": [
                "channel_id",
                "channel_name",
                "seller_id",
                "state"
            ],
            "properties": {
                "channel_id": {
                    "description": "渠道商",
                    "type": "integer"
                },
                "channel_name": {
                    "description": "渠道商名称",
                    "type": "string"
                },
                "memo": {
                    "description": "备注",
                    "type": "string"
                },
                "seller_id": {
                    "description": "运营商",
                    "type": "integer"
                },
                "state": {
                    "description": "状态 1开启,2关闭",
                    "type": "integer"
                }
            }
        },
...
        "service_admin.DeleteXChannelReq": {
            "type": "object",
            "required": [
                "channel_id",
                "seller_id"
            ],
            "properties": {
                "channel_id": {
                    "description": "渠道商",
                    "type": "integer"
                },
                "seller_id": {
                    "description": "运营商",
                    "type": "integer"
                }
            }
        },
...
        "service_admin.GetXConfigReq": {
            "type": "object",
            "required": [
                "seller_id"
            ],
            "properties": {
                "channel_id": {
                    "description": "渠道商",
                    "type": "integer"
                },
                "config_name": {
                    "description": "配置名称",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "seller_id": {
                    "description": "运营商",
                    "type": "integer"
                }
            }
        },
        "service_admin.GetXConfigRes": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.XConfig"
                    }
                }
            }
        },
...
        "service_admin.UpdateXChannelReq": {
            "type": "object",
            "required": [
                "channel_id",
                "channel_name",
                "seller_id",
                "state"
            ],
            "properties": {
                "channel_id": {
                    "description": "渠道商",
                    "type": "integer"
                },
                "channel_name": {
                    "description": "渠道商名称",
                    "type": "string"
                },
                "memo": {
                    "description": "备注",
                    "type": "string"
                },
                "seller_id": {
                    "description": "运营商",
                    "type": "integer"
                },
                "state": {
                    "description": "状态 1开启,2关闭",
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "": {
            "type": "apiKey",
            "name": "x-token",
            "in": "header"
        }
    }
}